+++ models/Produto.js
@@ -4,7 +4,7 @@ module.exports = (connection, DataTypes) => {
 
     const model = connection.define('Produto', {
          id:{
-             type: DataTypes.INTEGER(100),
+             type: DataTypes.INTEGER,
              primaryKey: true,
              autoIncrement: true
          },
